Rename zmq[-_]app to zmq
