+++ Dockerfile
@@ -16,7 +16,7 @@ RUN python3 -m pip install . && \
   python3 -m pip install plugins/apps/threatbus_cif3 && \
   python3 -m pip install plugins/apps/threatbus_misp[zmq] && \
   python3 -m pip install plugins/apps/threatbus_zeek && \
-  python3 -m pip install plugins/apps/threatbus_zmq_app && \
+  python3 -m pip install plugins/apps/threatbus_zmq && \
   python3 -m pip install plugins/backbones/threatbus_inmem && \
   python3 -m pip install plugins/backbones/threatbus_rabbitmq
 COPY config* ./
