+++ plugins/apps/threatbus_zmq/setup.py
@@ -24,7 +24,7 @@ setup(
         "Topic :: System :: Distributed Computing",
     ],
     description="A plugin to connect apps via ZeroMQ.",
-    entry_points={"threatbus.app": ["zmq-app = threatbus_zmq_app.plugin"]},
+    entry_points={"threatbus.app": ["zmq = threatbus_zmq.plugin"]},
     install_requires=[
         "pyzmq>=19",
         "python-dateutil>=2.8.1",
@@ -43,8 +43,8 @@ setup(
     license="BSD 3-clause",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    name="threatbus-zmq-app",
-    packages=["threatbus_zmq_app"],
+    name="threatbus-zmq",
+    packages=["threatbus_zmq"],
     python_requires=">=3.7",
     url="https://github.com/tenzir/threatbus",
     version="2021.07.29",
